+++ test/frontend/toolkit/spec/index.js
@@ -1,13 +1,12 @@
 /* eslint-disable max-len */
-const $ = require('jquery');
 
 describe('frontend/toolkit', function () {
   beforeAll(function () {
-    $(document.body).append($('<div id="test-container"/>'));
+    document.body.insertAdjacentHTML('beforeend', '<div id="test-container"/>');
   });
 
   beforeEach(function () {
-    $('#test-container').empty();
+    document.querySelector('#test-container').replaceChildren();
   });
 
   require('./helpers');